Signing All Artifacts
The following adds one block to sign all artifacts but also supplies the necessary information to generate these signatures. These are used when pulling from mavens to validate that the signature is from the original author. They can also be supplied with the download file along with the md5 and sha1 to verify that the jar information hasn't been tampered with.